Problem discovered!
When NPM installs a GitHub dependency, and that dependency is large
enough to have a tarball generated, the tarball is extracted into the
NPM cache directory, and then processing of the lifecycle scripts occurs
there. That context shift overwrites the INIT_CWD environment variable
with the cache path, thereby losing access to the project folder.

For now, Antora Doc Tools needs to be cloned, and then projects that
depend on it need to `npm i file:path/to/antora-doc-tools`.

Also fixed: use the correct path when updating ADT dictionary symlinks.